Rename FilePlaceholders to YamlPlaceholders to make the placeholder format explicit
